requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
pytest
slowapi
redis
cloudinary
razorpay
stripe
fastapi-cache2
twilio
sendgrid
geopy
haversine
requests
structlog
sentry-sdk[fastapi]